requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cached-property==2.0.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
DateTime==5.5
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.1.0
Keras==3.3.3
Keras-Preprocessing==1.1.2
Markdown==3.6
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.4.0
pyspark==3.5.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.13.0
six==1.16.0
tensorboard==2.16.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.15.0
termcolor==2.4.0
typing-extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wincertstore==0.2
wrapt==1.16.0
zipp==3.18.1
zope.interface==6.4
